moved settings to pages
