+++ src/pages/settings/Settings.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react"
 import { Link } from "react-router-dom"; //TODO add "Navigate" when needed
-import BoxContainer from "../globalComponents/BoxContainer";
-import Navbar from "../globalComponents/Navbar";
-import Inbox from "../rightComponent/Inbox";
+import BoxContainer from "../../components/globalComponents/BoxContainer";
+import Navbar from "../../components/globalComponents/Navbar";
+import Inbox from "../../components/rightComponent/Inbox";
 import DeleteAccount from "./DeleteAccount";
 import EditAccount from "./EditAccount";
